refactor: flatten API structure into core/ and estimators/
  - Remove api/ folder nesting
  - Split api/functions.py into core/{branch,logits,discrete}.py
  - Split api/estimators.py into estimators/{finite_difference,reinforce,gumbel_softmax}.py
  - Update all imports: mellowgate.api.* -> mellowgate.*
