--- src/mellowgate/api/__init__.py
@@ -1,28 +1,0 @@
-"""API module for mellowgate discrete optimization functions."""
-
-from .estimators import (
-    FiniteDifferenceConfig,
-    GumbelSoftmaxConfig,
-    ReinforceConfig,
-    ReinforceState,
-    finite_difference_gradient,
-    gumbel_softmax_gradient,
-    reinforce_gradient,
-)
-from .experiments import Sweep, run_parameter_sweep
-from .functions import Branch, DiscreteProblem, LogitsModel
-
-__all__ = [
-    "Branch",
-    "DiscreteProblem",
-    "FiniteDifferenceConfig",
-    "GumbelSoftmaxConfig",
-    "LogitsModel",
-    "ReinforceConfig",
-    "ReinforceState",
-    "Sweep",
-    "finite_difference_gradient",
-    "gumbel_softmax_gradient",
-    "reinforce_gradient",
-    "run_parameter_sweep",
-]
