+++ src/mellowgate/core/__init__.py
@@ -1,0 +1,12 @@
+"""Core API components for discrete optimization problems."""
+
+from .branch import Bound, Branch
+from .discrete import DiscreteProblem
+from .logits import LogitsModel
+
+__all__ = [
+    "Bound",
+    "Branch",
+    "DiscreteProblem",
+    "LogitsModel",
+]
